Add support for Ubuntu Base.
Install the base system from a downloadable image, then patch it with
the PACKAGES we need. This method produces a much smaller image than
with debootstrap, and the building process completes much more quickly.

The process also does not require root (with some caveats). The debian
program 'fakeroot' replaces all uses of 'sudo' to ensure that the user
and group of files in the final tarball remains the same.

Signed-off-by: Alistair Strachan <astrachan@google.com>