Add support for rootless prepare.
Support using 'debootstrap --foreign' instead of qemu-debootstrap and
eliminate all use of sudo. Use fakeroot to preserve permissions in the
output filesystem. The finalization of the debootstrap is done on the
device side.

This change enables the use of distributions which are not compatible
with the distribution debootstrap is being run on (normally not a
supported configuration by debootstrap without --foreign), including
alternative architecture and, more importantly, different glibc
versions.

Signed-off-by: Alistair Strachan <astrachan@google.com>