adeb: add prebuilt headers from Android common kernel
One of the biggest hurdles for new users is the need for kernel
source code / headers. This is an attempt to provide a working solution
(albeit not optimal) by installing prebuilt headers generated by
compiling the Android common kernel for arm64 and x86_64 architectures.

The headers were obtained form the public repository:
https://android.googlesource.com/kernel/common/

Supported versions are 4.9 and 4.14

Signed-off-by: Erick Reyes <erickreyes@google.com>