Fix Qemu image building
The --build-image flag can be used to build images for Qemu. The
download_headers function breaks this because it needs to query the
kernel version from the device, which is not possible with
"--build-image" since there may be no adb-connected device. Just
prevent the download_headers from triggering on Qemu images for now.

Signed-off-by: Joel Fernandes (Google) <joel@joelfernandes.org>