buildstrap: execute build-bcc.sh on the host using chroot
When chrooting into a debian image that was created for a different arch
the system will look like it's executing on that arch and we can do the
compilation directly from this env as if we were running natively on
that arch. Debian recognizes the arch difference and might be using qemu
directly behind the scene. Not entirely sure on the exact mechanisms
deployed here but have used this trick several times in the past.

Also fix a typo s/be/we/

Signed-off-by: Qais Yousef <qais.yousef@arm.com>