addons: improve scripts to run on busybox systems
The path in busybox or non-android system is different. So add two paths
keepingn the busybox like one at top as they don't deal with multiple
path names nicely.

Also on busybox like systems tar is not always capable and can't handle
untarring tgz in one go. But gunzip is usually available so break it
down into 2 steps.

Remove die as it's not supported on busybox like systems too.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>