androdeb: allow buildtar without flashing device
It's more convenient to have the buildtar step SKIP_DEVICE to generate
the tar then use the --archive option in the second step to do the
actual androdeb setup of the device.

Also since INSTALL_BCC is now performed when creating the debian image
we can always tar the deb.tar.gz without having to copy anything from
the device.

Also drop creating a zip version of the generated tarred file.

Signed-off-by: Qais Yousef <qais.yousef@arm.com>